+++ inc/operable.h
@@ -102,7 +102,7 @@ public:
 
   static bool cpu_side_reset_ready;
 
-  static uint8_t currently_active_thread_ID;
+  static uint16_t currently_active_thread_ID;
 
   static std::vector<uint64_t> reset_ins_count_global;
   // WL
